+++ library/src/main/java/com/chuckerteam/chucker/internal/support/JsonConverter.kt
@@ -5,11 +5,16 @@ import com.google.gson.GsonBuilder
 
 internal object JsonConverter {
 
-    val instance: Gson by lazy {
+    val nonNullSerializerInstance: Gson by lazy {
         GsonBuilder()
             .disableHtmlEscaping()
-            .serializeNulls()
             .setPrettyPrinting()
+            .create()
+    }
+
+    val instance: Gson by lazy {
+        nonNullSerializerInstance.newBuilder()
+            .serializeNulls()
             .create()
     }
 }