+++ library/src/main/java/com/chuckerteam/chucker/internal/support/RequestProcessor.kt
@@ -25,6 +25,7 @@ internal class RequestProcessor(
 
     private fun processMetadata(request: Request, transaction: HttpTransaction) {
         transaction.apply {
+            requestHeadersSize = request.headers.byteCount()
             setRequestHeaders(request.headers.redact(headersToRedact))
             populateUrl(request.url)