+++ library/src/main/java/com/chuckerteam/chucker/internal/support/TransactionDetailsHarSharable.kt
@@ -1,0 +1,11 @@
+package com.chuckerteam.chucker.internal.support
+
+import android.content.Context
+import okio.Buffer
+import okio.Source
+
+internal class TransactionDetailsHarSharable(
+    private val content: String,
+) : Sharable {
+    override fun toSharableContent(context: Context): Source = Buffer().writeUtf8("$content\n")
+}