+++ library/src/main/java/com/chuckerteam/chucker/internal/ui/transaction/TransactionActivity.kt
@@ -12,8 +12,10 @@ import androidx.viewpager.widget.ViewPager
 import com.chuckerteam.chucker.R
 import com.chuckerteam.chucker.databinding.ChuckerActivityTransactionBinding
 import com.chuckerteam.chucker.internal.data.entity.HttpTransaction
+import com.chuckerteam.chucker.internal.support.HarUtils
 import com.chuckerteam.chucker.internal.support.Sharable
 import com.chuckerteam.chucker.internal.support.TransactionCurlCommandSharable
+import com.chuckerteam.chucker.internal.support.TransactionDetailsHarSharable
 import com.chuckerteam.chucker.internal.support.TransactionDetailsSharable
 import com.chuckerteam.chucker.internal.support.shareAsFile
 import com.chuckerteam.chucker.internal.support.shareAsUtf8Text
@@ -80,9 +82,18 @@ internal class TransactionActivity : BaseChuckerActivity() {
         R.id.share_curl -> shareTransactionAsText { transaction ->
             TransactionCurlCommandSharable(transaction)
         }
-        R.id.share_file -> shareTransactionAsFile { transaction ->
+        R.id.share_file -> shareTransactionAsFile(EXPORT_TXT_FILE_NAME) { transaction ->
             val encodeUrls = viewModel.encodeUrl.value!!
             TransactionDetailsSharable(transaction, encodeUrls)
+        }
+        R.id.share_har -> shareTransactionAsFile(EXPORT_HAR_FILE_NAME) { transaction ->
+            TransactionDetailsHarSharable(
+                HarUtils.harStringFromTransactions(
+                    listOf(transaction),
+                    getString(R.string.chucker_name),
+                    getString(R.string.chucker_version)
+                )
+            )
         }
         else -> super.onOptionsItemSelected(item)
     }
@@ -105,24 +116,26 @@ internal class TransactionActivity : BaseChuckerActivity() {
         return true
     }
 
-    private fun shareTransactionAsFile(block: (HttpTransaction) -> Sharable): Boolean {
-        val transaction = viewModel.transaction.value
-        if (transaction == null) {
-            showToast(getString(R.string.chucker_request_not_ready))
-            return true
-        }
-
-        val sharable = block(transaction)
+    private fun shareTransactionAsFile(fileName: String, block: suspend (HttpTransaction) -> Sharable): Boolean {
         lifecycleScope.launch {
-            val shareIntent = sharable.shareAsFile(
-                activity = this@TransactionActivity,
-                fileName = EXPORT_FILE_NAME,
-                intentTitle = getString(R.string.chucker_share_transaction_title),
-                intentSubject = getString(R.string.chucker_share_transaction_subject),
-                clipDataLabel = "transaction"
-            )
-            if (shareIntent != null) {
-                startActivity(shareIntent)
+            val transaction = viewModel.transaction.value
+            if (transaction == null) {
+                showToast(getString(R.string.chucker_request_not_ready))
+                return@launch
+            }
+
+            val sharable = block(transaction)
+            lifecycleScope.launch {
+                val shareIntent = sharable.shareAsFile(
+                    activity = this@TransactionActivity,
+                    fileName = fileName,
+                    intentTitle = getString(R.string.chucker_share_transaction_title),
+                    intentSubject = getString(R.string.chucker_share_transaction_subject),
+                    clipDataLabel = "transaction"
+                )
+                if (shareIntent != null) {
+                    startActivity(shareIntent)
+                }
             }
         }
         return true
@@ -141,7 +154,8 @@ internal class TransactionActivity : BaseChuckerActivity() {
     }
 
     companion object {
-        private const val EXPORT_FILE_NAME = "transaction.txt"
+        private const val EXPORT_TXT_FILE_NAME = "transaction.txt"
+        private const val EXPORT_HAR_FILE_NAME = "transaction.har"
         private const val EXTRA_TRANSACTION_ID = "transaction_id"
         private var selectedTabPosition = 0